+++ android/src/main/java/dev/barooni/capacitor/calendar/models/inputs/GetDefaultCalendarInput.kt
@@ -5,5 +5,5 @@ import com.getcapacitor.PluginCall
 data class GetDefaultCalendarInput(
     private val call: PluginCall,
 ) {
-    val useFallbackCalendar: Boolean = call.getBoolean("useFallbackCalendar", false)
+    val useFallbackCalendar: Boolean = call.getBoolean("useFallbackCalendar", false) ?: false
 }