+++ src/index.ts
@@ -34,6 +34,7 @@ import type { EventGuest } from './schemas/interfaces/event-guest';
 import type { GetDefaultCalendarOptions } from './schemas/interfaces/get-default-calendar-options';
 import type { GetReminderByIdOptions } from './schemas/interfaces/get-reminder-by-id-options';
 import type { GetRemindersFromListsOptions } from './schemas/interfaces/get-reminders-from-lists-options';
+import type { ListCalendarsResult } from './schemas/interfaces/list-calendars-result';
 import type { ListEventsInRangeOptions } from './schemas/interfaces/list-events-in-range-options';
 import type { ModifyCalendarOptions } from './schemas/interfaces/modify-calendar-options';
 import type { ModifyEventOptions } from './schemas/interfaces/modify-event-options';
@@ -87,6 +88,7 @@ export type {
   GetDefaultCalendarOptions,
   GetReminderByIdOptions,
   GetRemindersFromListsOptions,
+  ListCalendarsResult,
   ListEventsInRangeOptions,
   ModifyCalendarOptions,
   ModifyEventOptions,